% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import_latest_tweets.R,
%   R/fct_import_tweets.R
\name{import_latest_tweets}
\alias{import_latest_tweets}
\title{Import latest tweets}
\usage{
import_latest_tweets(n)

import_latest_tweets(n)
}
\arguments{
\item{n}{}
}
\value{



}
\description{
Import latest tweets

Title
}
