% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_install_py_dependencies.R
\name{install_py_dependencies}
\alias{install_py_dependencies}
\title{install_py_dependencies}
\usage{
install_py_dependencies(method = "auto", conda = "auto")
}
\arguments{
\item{conda}{}
}
\value{

}
\description{
Wrapper function that can be used to conveniently install all Python dependencies.
}
