% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(use_venv = T, ...)
}
\arguments{
\item{use_venv}{Boolean: Should virtual environment be set up when running from local?}

\item{...}{A series of options to be used inside the app.}
}
\description{
Run the Shiny Application
}
